bugfix: some property-based filter were incorrectly parsed
This usually lead to a syntax error on startup and rsyslogd not actually
starting up. The problem was the regex, which did not care for double
quote characters to follow in the action part - unfortunately something
that can frequently happen with v6+ format. An example:
:programname, isequal, "as" {action(type="omfile" ...) }
Here, the part
:programname, isequal, "as" {action(type="omfile"
was treated as the property filter, and the rest as action part.
Obviously, this did not work out. Unfortunately, such situations usually
resulted in very hard to understand error messages.