optimze: reduce memory operations during dns resolution/hostname setting
previously, hostname and ip strings were shuffled to the msg object, which
created a property out of them. Now the cache holds the property, and it
is resused (almost) everywhere, what saves a lot of memory operations.

The only exception is imtcp session setup, where different handling
of the hostname is done, which we need to sort out (but that's another
story).