logsig: milestone/verfier: block timestamp verification almost complete
unfortunately, there seems to be a problem with the GuardTime API, so
that I need their support before being able to carry on. Once I
receive it, it should be fairly quick to complete the function. I am
commiting this work as I do not know how long it will take to receive
an answer.