provide support for user-settable shutdown indicator
This was introduced in support of rsyslog, but is probably useful
for other users as well. In short, it provides the ability to set
a pointer that is expected to be zero as long as the library shall
continue to run and set to 1 if it shall terminate. That ptr is then
checked at appropriate places of the library code.