bugfix: postpone the start of the imgssapi listener
If the listen socket is opened earlier (during configuration parsing),
it is closed again during early init.

The start has been postponed to the 'runInput' stage, which has an
unpleasant side effect that it will fail if the priviledges are
dropped.

The code should be eventually fixed to support the new config system.