systemd: use service type notify
There is a time window, between rsyslog reporting syntax
errors and the daemon returning with failure, this may cause

systemctl restart rsyslog to not report any error inmediately
but later in the logs which is confusing to users.

The appropiate steps to correct this annoyance is to notify
systemd with a simple sd_notify(0, "READY=1"); just before
entering the main loop.

Tested in openSUSE 12.3/13.1 x86_64