bugfix: prevent a segfault during imtcp's shutdown
If imtcp is loaded and no listeners are configured (which is uncommon),
rsyslog crashes during shutdown.

The cause is that startWorkerPool() is called unconditionally during
modExit() even if there are no listerners. Correcting this triggers a
memleak of the wrkrMut mutex which shouldn't have been destroyed in
startWorkerPool() in the first place.