Add pom to package easily the java part with maven
Maven will create a target dir with its compilation files and three jars,
one with compiled classes, one with classes source code and one with
javadocs.

The generated jar is not executable directly. For instance, we will
have to do:
    java -cp target/rsyslog-8.1.6.jar com.rsyslog.gui.diaggui.DiagGUI
to execute the main in com.rsyslog.gui.diaggui.DiagGUI class.

The previous build system with makefile is not affected by this patch.