add "edit on github" to doc sidebar
based on the recipe found here:
   https://gist.github.com/mgedmin/6052926#file-edit_on_github-pyb

Note that it does not yet work correctly, but it looks like we
can tweak it to make it do what we want. I am just committing so
that it is clear what we have taken from the github gist and what
we have modified ourselves.