Patch for bug in omzmq plugin
I had an issue with rsyslog backing up and consuming tons of memory if there wasn't a zmq socket listening for messages when using the `omzmq3` plugin. Setting the `sendHWM` option was not working because zstr_send blocks when the HWM (High water mark) is hit, causing the buffer on rsyslog to continue to fill up. Switching to `zmq_send` and specifying ZMQ_NOBLOCK does the trick. Now settings HWM and other options actually work.