remove contribution from mono_matsuko
we could not reach him to agree to the re-licensing to ASL 2.0.
So I remove his contribution, which means we do no longer need
has agreement. This was the original contribution:

http://git.adiscon.com/?p=rsyslog.git;a=commitdiff;h=fb7fab511e0de2566c0de47468b443ef5bb508d3

Note that the bug is now introduced again. We will fix it via a
different method in the next commit.