fix broken PostgreSQL database creation
Quoting is really really important under PostgreSQL. Giving the
database name in single quotes simply won't work. Giving it in double
quotes is not the same as not quoting it at all. rsyslog has
traditionally connected case-sensitively to "Syslog" but accessed the
tables case-insensitively so we should follow suit here.

Furthermore, you cannot create a table based on template1 with a
different encoding. template0 must be used instead.