bugfix: omelasticsearch does not work with broken/changed ES 1.0+ API
closes: https://github.com/rsyslog/rsyslog/issues/104

IMPORTANT: This is untested, as setting up the env is quite some pain
for me. I push this in order to get feedback from a couple of folks that
have committed to test out the patch. Do NOT consider this fixed until
it has been merged to the official "v8-stable" (or above) branch.