imfile: refactor "files" array to use doubly-linked list
The array did not support dynamic file monitors as they happen
when wildcards are supported. While of course we could have found
a way to make this work with the array, this would not have been
a proper solution - much more complex than needed. So we solve
this by using a doubly linked list.