bugfix (kind of): write pid file earlier if privileges are dropped
We moved writing the pid file to after the initialization, so that
it is written only if the rsyslog process actually runs. However,
when dropping privileges, this can cause problems when rsyslog then
no longer has sufficient permission to write the pid file. This
happens frequently, as the pid file is usually stored directly under
/var/run.

A proper solution would be to create a rsyslog-user-writable directory
under /var/run, and make other scripts use that. However, this causes
grief for a lot of distros, and so we create a work-around: when
privileges are dropped, the pid file is again written a bit too early,
before privileges are dropped. This can lead to a false sense of
startup, but given the fact that this was no big problem the past
10+ years, it isn't supposed to be a big one for the coming years.

If no privileges are dropped, the pid file is still written at the
"correct" place, when rsyslog has completely been started up.

closes https://github.com/rsyslog/rsyslog/issues/169