omkafka: add stats and dynamic topic
stats:
  - submitted: producer submitted message to broker.
  - failures: producer failed to submit message to broker.

Topic is able to be a template, the topics are stored in the instanceData and the kafka_rd_t object has been moved into instanceData. This is alright because librdkafka is threadsafe and to acheive the dynamic topic results it is required.
The topics are stored in a omfile like dynamic cache due to librdkafka leaking memory when passing in a conf object, plus us keeping a local topic "list" makes the subsequent topic conf and rd_kafka_topic_new calls unnecessary. The move of the topic down the writeKafka is for dynamic parsing of the topic for each message that comes through.

Additional items:
  - Added safety to rd_kafka_t destruction, where it flushes the queued messages before object destruction.
  - Dynamic topic is enabled with dynatopic="on" in the configuration.
  - Dynamic topic LRU cache limit can be set with dynatopic.cachesize configuration option
  - On HUP the active topics are closed and free'd

These changes have been run through valgrind and I can confirm there is no leaking (at least within my setup).