testebench: mitigate UDP message loss
in legacy config, we do not have a way to slow down sending
UDP messages to a rate where no message loss occurs (message
loss is fine here due to the UDP RFCs!). As such, we simply
reduce the number of messages in this test. That actually
doesn't hurt anything, as the purpose of the test is that
messages are properly received and processed.

More messages are still being sent with the equivalent test
that uses RainerScript config.