Make configure fail when building with "--enable-omkafka" but librdkafka wasn't found
We were only checking for headers but haven't defined an "if-not-found" action.
So when building with "--enable-omkafka" but without librdkafka installed,
compilation would fail.

Now we are checking for librdkafak first via PKG_CHECK_MODULES but will fall back to
AC_CHECK_HEADER for librdkafka packages not providing a .pc file.
If these checks will fail, we will now show a friendly error message.