bugfix: large memory consumption on suspended action
When an action is suspended, rsyslog's memory consumption continues
to grow. On very busy systems, this can lead to system stability
errors as it is possible that rsyslog allocates too much memory.

The root cause of this issue is the invalid management of a string
cache index, which is reset to the beginning in the wrong function,
one that is not reached when suspended. This in turn leads to
continous increase of the cache.

This fix moves the reset to the proper location.

Note that this problem was not detectable by our valgrind tests
because proper cleanup was done on rsyslog termination (but only
then).

closes https://github.com/rsyslog/rsyslog/issues/253