bugfix: size-based legacy config statements did not work properly
on some platforms they misadressed memory, which could also lead
to a segfault on startup. The problem is NOT experience on amd64
in 64bit builds -- that's probably the reason this bug was uncovered
very late. We assume, it's present in all v8 versions.

Thanks to Michael Biebl for alerting us of it.

closes https://github.com/rsyslog/rsyslog/issues/270