Make llDestroy safe
Keep the destructed list in a consistent state as the provided
destructor may access the very same list again before llDestroy
finishes.

The previous approach could actually trigger a segmentation violation
error for specific scenarios and configurations.
In one particular case, destructing the list of rulesets lead to
shutdown of an action DA queue and spooling its messages to disk which
in turn triggered a search for the ruleset name of the message trought
the destructed list.

The change probably slightly degrades performace, but that shouldn't
have an observable effect in the current code base.