Fix destruction of modules without cfSysLineHdlrs
Modules that don't register cfSysLineHdlrs were not completely
destructed, which resulted in a mem leak. Most of the time, this is
not an issue, as the majority of modules is destructed only during
shutdown. In case of modules which can be (un)loaded repeatedly, such
as lmnsd_ptcp, the leaked amount would accumulate over time. Still,
overall, the leak is pretty small.