bugfix: fix default action naming
The default name generated for an action references a global counter
instead of the action-specific number. In practice this means that the
number in the name will be greater by at least one than the assined
number most of the time.

NB: The code generating various action-related messages is using the
action name inconsistently; sometimes the generated name is used,
sometimes the action number is used, thus producing "action N" or
"action N+1" identifiers at different locations in the code.