bugfix: imudp could do 0-size calloc if no listeners are active at late stage
This was found by clang static analyzer and is extremely unlikely to
happen in practice, due to checks in earlier stages. But we did not
investigate that in detail but rather add checks to be on the safe
side. Note that this is not performance critical as it happens just
on startup.