fix ommysql issues under RHEL7
under RHEL 7, we have an issue when gnutls is linked. For some reason
(autoconf magic?) we link ommysql against gnutls. I have not found the
root cause, but setting LIBS= below makes the problem go away. A better
solution would be appreciated, but for the time being this seems to
solve the issue.
closes https://github.com/rsyslog/rsyslog/issues/408