imfile: Fixed file table corruption when monitoring single files
When monitoring single files in inotify mode, a corruption of the
configuration filetable happened when the file was deleted.
The Active and Configuration table are sharing the same Listener
object when monitoring a single file. We check if the listener
object is present in the configured filetable before we delete it now.

closes https://github.com/rsyslog/rsyslog/issues/392