This fixes a bug which allowed omkafka to get into a state where it wouldn't accept any new messages for publishing.
It also fixes a bug which whould manifest only when using dyntopic and when producer is in i-won't-accept-new-messages state, and caller tries to produce a new message. The dyn-topic-lock wouldn't be unlocked (because of premature fail-jump).

If you know librdkafka at a high level and are interested in details, here is a more detailed explaination:
So 'rk' tries to bound number of objects in flight by maintaining a counter, the max-value of which defaults to 100k. When a new message comes in, it rejects it if it already is at max-messages-in-flight. Now this counter is decremented (and in-flight-msg memory is freed) in callback-poll (which picks things from reply-queue, rd_kafka_op_t for delivery-ack op). If we don't poll when we are already at max-in-flight-messages and there is no poll in flight, it'll basically drain all 100k in-flight message and then start idling (because rk->rk_u.producer->msg_cnt still thinks it has 100k messages in flight, and delivery-acks are yet to be handled, but of-course no one handles them).