diskqueue: Fixed two minor issues when syncqueuefiles is enabled.
First issue: Error 14 was generated on the .qi file directory handle.
As the .qi filestream does not have a directory set, fsync
was called on an empty directory causing a error 14 in debug log.
closes https://github.com/rsyslog/rsyslog/issues/402

Second issue: When queue files existed on startup, the bSyncQueueFiles
strm property was not set to 1. This is now done in the
qqueueLoadPersStrmInfoFixup function.
closes https://github.com/rsyslog/rsyslog/issues/403