testbench: Fix potential buffer overflow detection
On some architectures, calculation of edLen may lead to a
negative value when the value returned by rand() is less
than extraDataLen away from RAND_MAX. The negative value
passed to memset() triggers a glibc error. Forcing the
cast of rand() to be unsigned avoid the return of a negative
value.

Closes: #506

Signed-off-by: Louis Bouchard <louis.bouchard@canonical.com>