remove unofficial lookup table code
This code officially never existed and was, in its incomplete
state, added as a convenience to some users.

Background (from ML posting):
Sorry, folks, good intent always seems to find someone who turns it
into negative. I was yesterday contacted by a distro maintainer who
wants to turn this bug in the officially non-existant lookup table
feature into a CVE and insists that it is a vuln even after the
argument that the feature never oficially existed.

It looks like it was a bad idea to merge potentially useful yet
incomplete code into the main branch (and documenting it to be not
present). It looks like I need to re-think my stance on experimental
features.

Anyhow, I really don't want to support the argument that something
non-existing can be a CVE. As such, I will create a new
master-insecure branch, which will be a clone of the current master
branch. Then I'll remove the lookup table code, so that the code base
matches the documentation. I really don't want to create a general
principle here that we need to create CVEs (and patched) for something
that was just added as a convenience for a handful of folks who were
ready to take a risk.

If there is sufficient interest, we can consider officially adding
this feature to the January 8.15 release iff it is ready by then.
@janmejay: please let me know if you would like to continue with your
work on lookup tables under this new situation.

As soon as I have time, I'll check what else needs to be removed. Not
sure about the ./contributed branch, because the project cannot
guarantee at all this is bug-free. It's documented to be so, but if
that is not sufficient, it should probably live only in the
master-insecure branch.