protect json variable access by read-write-lock
json-c objects cannot be used on multiple threads and so we
need to guard against this. Note that this was already handled
in many cases by totally locking the message, what we now do
no longer. So all in all, it could even improve performance for
configs where json is used. For others, it slightly decreases,
because we need to initialize the rwlocks for each message.