fix rawmsg-after-pri property
The index offAfterPRI inside the message object is supposed
to point to the first char in rawmsgs after PRI. Unfortunately,
it seems NOT to be correct/consistent in all cases. imuxsock
and imudp seem to have other values than imptcp. Testbench
covers some of that. As a work-around, we caluculate
the value ourselfes.