Ensure non-var property names are case-insensitive
We want to make sure that the defined message properties are always
compared case-insenstive, since this is the intention with these
properties today.

This should result it no actual changes to existing configuration file
behaviors.

In addition, we ensure the two property name/id mapping routines are
coded in the same order to help determine they are in-sync.